Core: separate indexing functions into own header
Indexing.hpp for now contains exactly what was in Vectormath.hpp before.
Next steps should be to rename the namespace and clean up the include
structure to reduce the places where Indexing.hpp is included.
